Fix module registration, article agreement and a float cast
Driving the interface in a real browser surfaced three bugs that unit-level
testing had missed.

DPG.use() only accepted a function, but timer.js, export.js and
demo-editor.js all register an object with init/onProject methods. Every one
was silently dropped, so the challenge timer, all three export buttons and
the Open demo button did nothing at all. It now takes either shape.

industries.json hard-codes "a {product}" in its background sentences, and 17
of the products start with a vowel sound, producing "a insurance plan" and
"a online booking service". One template hard-codes "an", which was equally
wrong for consonant products. The article can only be settled once the
product is in the sentence, so it is corrected after substitution, leaving
"a unisex range", "a used vehicle" and "a one-off" alone.

format_minutes() passed floor()'s float to _n(). WordPress compares the
plural form loosely so it was right in practice, but the cast makes it
correct regardless of the translation backend.

+++ design-project-generator/includes/class-export.php
@@ -187,7 +187,7 @@ class DPG_Export {
 			return sprintf( _n( '%d minute', '%d minutes', $minutes, 'design-project-generator' ), $minutes );
 		}
 
-		$hours = floor( $minutes / 60 );
+		$hours = (int) floor( $minutes / 60 );
 		$rest  = $minutes % 60;
 
 		if ( 0 === $rest ) {
